+++ lib/src/revset.rs
@@ -218,6 +218,7 @@ pub enum RevsetExpression {
         /// Copy of `repo.view().heads()`, should be set by `resolve_symbols()`.
         visible_heads: Option<Vec<CommitId>>,
     },
+    Coalesce(Rc<Self>, Rc<Self>),
     Present(Rc<Self>),
     NotIn(Rc<Self>),
     Union(Rc<Self>, Rc<Self>),
@@ -439,6 +440,20 @@ impl RevsetExpression {
         Rc::new(Self::Difference(self.clone(), other.clone()))
     }
 
+    /// Commits that are in the first expression in `expressions` that is not
+    /// `none()`.
+    pub fn coalesce(expressions: &[Rc<Self>]) -> Rc<Self> {
+        match expressions {
+            [] => Self::none(),
+            [expression] => expression.clone(),
+            _ => {
+                // Build balanced tree to minimize the recursion depth.
+                let (left, right) = expressions.split_at(expressions.len() / 2);
+                Rc::new(Self::Coalesce(Self::coalesce(left), Self::coalesce(right)))
+            }
+        }
+    }
+
     /// Resolve a programmatically created revset expression.
     ///
     /// In particular, the expression must not contain any symbols (bookmarks,
@@ -528,6 +543,7 @@ pub enum ResolvedExpression {
         candidates: Box<Self>,
         count: usize,
     },
+    Coalesce(Box<Self>, Box<Self>),
     Union(Box<Self>, Box<Self>),
     /// Intersects `candidates` with `predicate` by filtering.
     FilterWithin {
@@ -853,6 +869,14 @@ static BUILTIN_FUNCTION_MAP: Lazy<HashMap<&'static str, RevsetFunction>> = Lazy:
             visible_heads: None,
         }))
     });
+    map.insert("coalesce", |diagnostics, function, context| {
+        let ([], args) = function.expect_some_arguments()?;
+        let expressions: Vec<_> = args
+            .iter()
+            .map(|arg| lower_expression(diagnostics, arg, context))
+            .try_collect()?;
+        Ok(RevsetExpression::coalesce(&expressions))
+    });
     map
 });
 
@@ -1171,6 +1195,12 @@ fn try_transform_expression<E>(
                     visible_heads: visible_heads.clone(),
                 }
             }),
+            RevsetExpression::Coalesce(expression1, expression2) => transform_rec_pair(
+                (expression1, expression2),
+                pre,
+                post,
+            )?
+            .map(|(expression1, expression2)| RevsetExpression::Coalesce(expression1, expression2)),
             RevsetExpression::Present(candidates) => {
                 transform_rec(candidates, pre, post)?.map(RevsetExpression::Present)
             }
@@ -2050,6 +2080,10 @@ impl VisibilityResolutionContext<'_> {
                 let context = VisibilityResolutionContext { visible_heads };
                 context.resolve(candidates)
             }
+            RevsetExpression::Coalesce(expression1, expression2) => ResolvedExpression::Coalesce(
+                self.resolve(expression1).into(),
+                self.resolve(expression2).into(),
+            ),
             RevsetExpression::Present(_) => {
                 panic!("Expression '{expression:?}' should have been resolved by caller");
             }
@@ -2128,6 +2162,9 @@ impl VisibilityResolutionContext<'_> {
             RevsetExpression::AsFilter(candidates) => self.resolve_predicate(candidates),
             // Filters should be intersected with all() within the at-op repo.
             RevsetExpression::AtOperation { .. } => {
+                ResolvedPredicateExpression::Set(self.resolve(expression).into())
+            }
+            RevsetExpression::Coalesce(_, _) => {
                 ResolvedPredicateExpression::Set(self.resolve(expression).into())
             }
             RevsetExpression::Present(_) => {
@@ -2568,6 +2605,35 @@ mod tests {
             CommitRef(WorkingCopy(WorkspaceId("default"))),
         )
         "###);
+        insta::assert_debug_snapshot!(
+            RevsetExpression::coalesce(&[]),
+            @"None");
+        insta::assert_debug_snapshot!(
+            RevsetExpression::coalesce(&[current_wc.clone()]),
+            @r###"CommitRef(WorkingCopy(WorkspaceId("default")))"###);
+        insta::assert_debug_snapshot!(
+            RevsetExpression::coalesce(&[current_wc.clone(), foo_symbol.clone()]),
+            @r#"
+        Coalesce(
+            CommitRef(WorkingCopy(WorkspaceId("default"))),
+            CommitRef(Symbol("foo")),
+        )
+        "#);
+        insta::assert_debug_snapshot!(
+            RevsetExpression::coalesce(&[
+                current_wc.clone(),
+                foo_symbol.clone(),
+                bar_symbol.clone(),
+            ]),
+            @r#"
+        Coalesce(
+            CommitRef(WorkingCopy(WorkspaceId("default"))),
+            Coalesce(
+                CommitRef(Symbol("foo")),
+                CommitRef(Symbol("bar")),
+            ),
+        )
+        "#);
     }
 
     #[test]